requirements.txt
FastMCP
httpx
nba_api
anthropic
nest_asyncio
pypdf2